enhance use query hooks
